+++ oneflow/ir/include/OneFlow/OneFlowPatterns.td
@@ -160,57 +160,7 @@ def FusedPadConv2DPattern : Pat<
   ]
 >;
 
-def GetDefaultSeed :
-  NativeCodeCall<"mlir::oneflow::GetDefaultSeed($_builder)">;
-
-def FusedBiasAddMaskScale :
-  NativeCodeCall<"mlir::oneflow::CreateFusedBiasAddMaskScale($_builder, $0, $1, $2)">;
-
 def IsAddToOutputNone: Constraint<CPred<"mlir::oneflow::IsAddToOutputNone($0)">, "">;
-
-def FusedBiasAddDropoutPattern : Pattern<
-  (
-    OneFlow_DropoutOp: $dropout_res
-    (
-      OneFlow_BiasAddOp: $bias_add_res
-        $a,
-        $b,
-        $bias_add_op_name,
-        $bias_add_device_tag,
-        $bias_add_device_name,
-        $bias_add_scope_symbol_id,
-        $bias_add_hierarchy,
-        $bias_add_op_axis
-    ),
-    $_add_to_output,
-    $dropout_op_name,
-    $dropout_device_tag,
-    $dropout_device_name,
-    $dropout_scope_symbol_id,
-    $dropout_hierarchy,
-    $dropout_op_rate
-  ),
-  [
-    (
-      FusedBiasAddMaskScale
-      $dropout_res__0,
-      $bias_add_res,
-      (
-        OneFlow_RandomMaskLikeOp : $mask
-          $a,
-          $bias_add_op_name,
-          $dropout_device_tag,
-          $dropout_device_name,
-          $dropout_scope_symbol_id,
-          $dropout_hierarchy,
-          $dropout_op_rate,
-          (GetDefaultSeed)
-      )
-    ),
-    (replaceWithValue $mask)
-  ],
-  [(IsAddToOutputNone $_add_to_output)]
->;
 
 // TODO: use either to merge two patterns
 def FusedScaleTrilPattern2 : Pat<
